Migrated logic related to expection handler to respective files from main.py, reduced and removed the unused imports from main.py
